task

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,8 +8,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int x = n ^ m;
-	int f = 0;
+	unsigned long int x = n ^ m;
+	unsigned int f = 0;
 
 	while (x)
 	{
